# packages/core/src/libs/openapi/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vocion API",
    "version": "0.1.0",
    "description": "The HTTP API behind Vocion's dashboard. Every endpoint is scoped to one\norganization — the credential decides which, and there is no way to ask for\nanother one's records.\n\n**Authenticating.** Send `Authorization: Bearer vcn_live_…` with a tenant API\ntoken, issued under Developers → API tokens. Calls made from a signed-in\nbrowser session are accepted too, which is how the dashboard itself uses these\nendpoints. When an `Authorization` header is present it is the credential: a\nbad token is a 401 and never falls back to a cookie.\n\n**Errors.** Every failure returns the same envelope, whatever the status:\n\n```json\n{ \"error\": { \"code\": \"VALIDATION_FAILED\", \"message\": \"workerId is required\", \"details\": null } }\n```\n\n**Paging.** List endpoints take `limit` (default 50, maximum 200) and\n`offset`. Out-of-range values fall back to the defaults rather than erroring.\n\n**How this document is made.** It is generated from the route handlers\nthemselves — their paths, their doc comments, the error codes they emit — by\n`npm run openapi:generate`, and a test fails when it falls out of step with\nthe code. Request and response bodies are documented by their field names and\nthe endpoint's prose; they are not yet described field-by-field as schemas."
  },
  "servers": [
    {
      "url": "/",
      "description": "This deployment."
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "A tenant API token, issued under Developers → API tokens. Looks like `vcn_live_…`."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "A stable, machine-readable code such as VALIDATION_FAILED."
              },
              "message": {
                "type": "string",
                "description": "What went wrong, in English, for a developer reading a log."
              },
              "details": {
                "type": [
                  "object",
                  "null"
                ],
                "description": "Extra context when the endpoint has any; null otherwise."
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "get_notifications",
        "summary": "Your notifications in this workspace, newest first — what the bell and the notifications page read.",
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "The request was malformed or failed validation. Codes: `VALIDATION_FAILED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "The bearer token or session was missing or invalid. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Your notifications in this workspace, newest first — what the bell and the\nnotifications page read. Each carries its kind (and the kind's label), the\ntitle and body, the link it opens, the record it is about, whether it is\nread, and every channel's delivery state (`sent`, `pending` with the retry\nreason, `failed` with why, `skipped` with why). `unread` is the unread\ntotal; `nextBefore` pages further back.\nAuth: dashboard session, or a tenant token (its minter's notifications).\n\nQuery parameters:\n- `unread` — `1` for unread only.\n- `limit` — page size, 1–100 (default 30).\n- `before` — only notifications older than this id (from `nextBefore`).",
        "parameters": [
          {
            "name": "unread",
            "in": "query",
            "required": false,
            "description": "`1` for unread only.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "page size, 1–100 (default 30).",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "only notifications older than this id (from `nextBefore`).",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/push/devices": {
      "get": {
        "operationId": "get_push_devices",
        "summary": "The devices you asked to be notified on — browsers subscribed from the bell and iPhones whose app registered — newest first, with the last error a device repor…",
        "tags": [
          "push"
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "The bearer token or session was missing or invalid. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "The devices you asked to be notified on — browsers subscribed from the bell\nand iPhones whose app registered — newest first, with the last error a\ndevice reported.\nAuth: dashboard session, or a tenant token (its minter's devices)."
      },
      "post": {
        "operationId": "post_push_devices",
        "summary": "Register a device for push.",
        "tags": [
          "push"
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "The request was malformed or failed validation. Codes: `VALIDATION_FAILED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "The bearer token or session was missing or invalid. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Register a device for push. The iOS app sends `platform: ios`, the APNs\ndevice token as hex, its `bundleId` and `environment` (`production` for\nTestFlight and the App Store, `sandbox` for a build run from Xcode). A\nbrowser sends `platform: web` with its PushSubscription's `endpoint` and\n`keys` (`p256dh`, `auth`). `label` names the device in settings. The same\ntoken again refreshes it (and moves it to you). 201 with the device.\nAuth: dashboard session (the app signs in with it), or a tenant token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "A JSON object. This endpoint reads its fields dynamically, so they are described in the prose above.",
                "properties": {},
                "additionalProperties": true
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete_push_devices",
        "summary": "Stop notifying this browser: removes your web subscription by its endpoint (the browser knows its endpoint, not the row id).",
        "tags": [
          "push"
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "The request was malformed or failed validation. Codes: `VALIDATION_FAILED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "The bearer token or session was missing or invalid. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Stop notifying this browser: removes your web subscription by its\nendpoint (the browser knows its endpoint, not the row id). A device by id\nis `DELETE /api/v1/push/devices/:id`.\nAuth: dashboard session, or a tenant token (its minter's devices).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "The fields this endpoint reads.",
                "properties": {
                  "endpoint": {
                    "description": "Read by this endpoint; see the description for what it must contain."
                  }
                },
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/v1/push/devices/{id}": {
      "delete": {
        "operationId": "delete_push_devices_id",
        "summary": "Remove one of your devices; it gets no more push.",
        "tags": [
          "push"
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "The request was malformed or failed validation. Codes: `VALIDATION_FAILED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "The bearer token or session was missing or invalid. Codes: `UNAUTHORIZED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No such record in this organization. Codes: `NOT_FOUND`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Remove one of your devices; it gets no more push. Another person's device\nis a 404.\nAuth: dashboard session, or a tenant token (its minter's devices).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The id from the URL.",
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
  }
}